src/app/pyproject.toml
[project]
name = "weather-analytics"
version = "0.1.0"
description = "FastAPI сервис-агрегатор погоды из нескольких публичных API"
readme = "README.md"
authors = [
  { name = "Dmitry Erofeev" },
  { name = "Eduard Polyakov" },
  { name = "Maxim Piskaev" },
]
license = { text = "MIT" }
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.30.0",
  "httpx>=0.27.0",
  "pydantic>=2.7.0",
  "pydantic-settings>=2.2.0",
  "pymongo>=4.15.5",
  "python-dotenv>=1.2.1",
  "loguru>=0.7.3",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0.0",
  "pytest-anyio>=0.0.0",
  "ruff>=0.6.0",
  "black>=24.0.0",
  "mypy>=1.10.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.pytest.ini_options]
addopts = "-ra"
testpaths = ["tests"]

[tool.ruff]
line-length = 88
target-version = "py311"
src = ["app", "tests"]

[tool.ruff.lint]
select = ["E", "F", "I", "B"]
ignore = ["E501"]

[tool.black]
line-length = 88
target-version = ["py311"]
